Cambio de atributos Entity(ofertas) + Prueba de getEmpresas() que devulve un JSON

+++ Proyecto1DAM_DAW/src/main/java/com/example/demo/repository/EmpresaRepository.java
@@ -4,6 +4,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import com.example.demo.model.Empresa;
 
-public interface EmpresaRepository extends JpaRepository<Empresa, Long> {
+public interface EmpresaRepository extends JpaRepository<Empresa, Integer> {
 
 }
